Move current requests to homepage
